Allows page title to be a string or array
Extends the `pageTitle` prop to accept either a string or an array of strings.

This change provides more flexibility in defining page titles, especially
when titles are dynamically constructed from multiple parts. It also ensures
that the title is properly formatted, filtering out any non-string parts
and joining the remaining parts with a space.

Addresses issue where the page title was not correctly rendered when
composed of multiple parts.

Related to #51